+++ src/config/common/mod.rs
@@ -1,0 +1,4 @@
+//! Common functionality for the configuration module.
+
+pub mod decoders;
+pub mod encoders;